Refactored names and split into more multiple smaller files
